+++ src/Pim/Bundle/BaseConnectorBundle/Processor/Denormalization/ArrayConverter/Structured/AttributeOptionToStandardConverter.php
@@ -2,11 +2,10 @@
 
 namespace Pim\Bundle\BaseConnectorBundle\Processor\Denormalization\ArrayConverter\Structured;
 
-use Pim\Bundle\BaseConnectorBundle\Exception\ArrayConversionException;
 use Pim\Bundle\BaseConnectorBundle\Processor\Denormalization\ArrayConverter\StandardArrayConverterInterface;
 
 /**
- * Convert structured format to standard format for attribute option
+ * Attribute Option Structured Converter
  *
  * @author    Nicolas Dupont <nicola@akeneo.com>
  * @copyright 2015 Akeneo SAS (http://www.akeneo.com)
@@ -15,8 +14,6 @@ use Pim\Bundle\BaseConnectorBundle\Processor\Denormalization\ArrayConverter\Stan
 class AttributeOptionToStandardConverter implements StandardArrayConverterInterface
 {
     /**
-     * {@inheritdoc}
-     *
      * Converts yaml array to standard structured array:
      *
      * Before:
@@ -42,53 +39,17 @@ class AttributeOptionToStandardConverter implements StandardArrayConverterInterf
      *         'fr_FR': '210 x 1219 mm'
      *     }
      * }
+     *
+     * @param array $item Representing a flat attribute option
+     *
+     * @return array structured product
      */
     public function convert(array $item)
     {
-        $this->validate($item);
+        // TODO: option resolver!
         $item['sort_order'] = $item['sortOrder'];
         unset($item['sortOrder']);
 
         return $item;
-    }
-
-    /**
-     * @param array $item
-     *
-     * @throws ArrayConversionException
-     */
-    protected function validate(array $item)
-    {
-        $requiredFields = ['attribute', 'code'];
-        foreach ($requiredFields as $requiredField) {
-            if (!in_array($requiredField, array_keys($item))) {
-                throw new ArrayConversionException(
-                    sprintf(
-                        'Field "%s" is expected, provided fields are "%s"',
-                        $requiredField,
-                        implode(', ', array_keys($item))
-                    )
-                );
-            }
-        }
-
-        $authorizedFields = array_merge($requiredFields, ['sortOrder', 'labels']);
-        foreach ($item as $field => $data) {
-            if (!in_array($field, $authorizedFields)) {
-                throw new ArrayConversionException(
-                    sprintf(
-                        'Field "%s" is provided, authorized fields are: "%s"',
-                        $field,
-                        implode(', ', $authorizedFields)
-                    )
-                );
-            }
-        }
-
-        if (isset($item['labels']) && !is_array($item['labels'])) {
-            throw new ArrayConversionException(
-                sprintf('Field "labels" must be an array, data provided is "%s"', print_r($item['labels'], true))
-            );
-        }
     }
 }
